requirements.txt
pytesseract>=0.3.10
Pillow>=9.0.0
keyboard>=0.13.5
pandas>=1.3.0
pyperclip>=1.8.2
